Put exit hook in extra file as it has to be used by more than one component
