+++ source/3dsexit.cpp
@@ -1,0 +1,15 @@
+#include <stdlib.h>
+#include <3ds.h>
+
+#include "3dsexit.h"
+
+void setExitFlag(APT_HookType hook, void* param)
+{
+    if (hook == APTHOOK_ONEXIT) {
+        appExiting = 1;
+    }
+}
+
+void enableExitHook() {
+    aptHook(&hookCookie, setExitFlag, NULL);
+}
